Исправлены баги, улучшение алгоритмов
- исправлен баг в отсутствии копирования текстурных координат при клонировании меша;
- в модуле перемещения точек ассерт заменен на выход из цикла;
- исправлена неправильная матрица трансформации при формировании эллипса, из-за чего было невозможно реконструировать наклонные объекты;
- более стабильный цикл формирования меша из слоев (учитывает ситуацию возможного зацикливания);
- входное  изобржение автоматически изменяет размер (при необходимости);
- добавлено несколько новых тестовых изображений.

+++ include/defs.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <limits>
 
-using Int = std::numeric_limits<double>;
+using Int = std::numeric_limits<int>;
 using Double = std::numeric_limits<double>;
 
 template<class T>
